requirements.txt
aiofiles==23.2.1
aiogram==3.7.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
certifi==2024.6.2
Django==5.0.6
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
pillow==10.3.0
psycopg2-binary==2.9.9
pydantic==2.7.3
pydantic_core==2.18.4
python-dotenv==1.0.1
SQLAlchemy==2.0.30
sqlparse==0.5.0
typing_extensions==4.12.1
yarl==1.9.4
